Refactor ToolObject class to use float for debuffAmount property
This commit updates the ToolObject class in the Assets/Scripts/Inventory Objects/ToolObject.cs file to change the type of the debuffAmount property from int to float. This change improves the accuracy and precision of the debuff amount values, making it more consistent with other properties in the class.

+++ Assets/Scripts/Inventory Objects/ToolObject.cs
@@ -31,7 +31,7 @@ public class ToolObject : ItemObject
     [SerializeField] float magPwr;
     [SerializeField] StatusType[] statusEffects;
     [SerializeField] Debuff[] debuffs;
-    [SerializeField] int debuffAmount;
+    [SerializeField] float debuffAmount;
 
     public void Awake()
     {
@@ -42,5 +42,5 @@ public class ToolObject : ItemObject
     public float MagPwr { get { return magPwr; } }
     public StatusType[] StatusEffects { get { return statusEffects; } }
     public Debuff[] Debuffs { get { return debuffs; } }
-    public int DebuffAmount { get { return debuffAmount; } }
+    public float DebuffAmount { get { return debuffAmount; } }
 }
